[feature#5] Add: managers.py
비즈니스 로직 views단에서 분리
